refactor: use internal interfaces rather than export them
