+++ status/batch/v1/register.go
@@ -1,7 +1,7 @@
 package v1
 
 import (
-	"github.com/caicloud/rudder-client/status/internal"
+	"github.com/caicloud/rudder-client/status/universal"
 	batchv1 "k8s.io/api/batch/v1"
 )
 
@@ -9,6 +9,6 @@ var (
 	gvkJob = batchv1.SchemeGroupVersion.WithKind("Job")
 )
 
-func Assist(u internal.Umpire) {
+func Assist(u universal.Umpire) {
 	u.Employ(gvkJob, JudgeJob)
 }
