+++ status/batch/v1beta1/register.go
@@ -1,7 +1,7 @@
 package v1beta1
 
 import (
-	"github.com/caicloud/rudder-client/status/internal"
+	"github.com/caicloud/rudder-client/status/universal"
 	batchv1beta1 "k8s.io/api/batch/v1beta1"
 )
 
@@ -9,6 +9,6 @@ var (
 	gvkCronJob = batchv1beta1.SchemeGroupVersion.WithKind("CronJob")
 )
 
-func Assist(u internal.Umpire) {
+func Assist(u universal.Umpire) {
 	u.Employ(gvkCronJob, JudgeCronJob)
 }
